PHP refactoring
