+++ Plugin/RenderingCacheContextPlugin.php
@@ -7,6 +7,12 @@ use Magento\Framework\App\Http\Context as HttpContext;
 use Magento\Framework\App\Request\Http;
 use Magento\Store\Model\StoreManagerInterface;
 
+/**
+ * The purpose of this class is to render different cached versions of the pages according to the user agent.
+ * If the "prevent backend rendering" configuration is turned on, we don't want to render the results on the backend
+ * side only for humans but we want to do it for the robots (configured in the "advanced" section of the extension).
+ * So with this plugin, two versions of the pages are cached : one for humans, and one for robots.
+ */
 class RenderingCacheContextPlugin
 {
     const RENDERING_CONTEXT = 'rendering_context';
